+++ src/skycoin/skycoin.go
@@ -349,6 +349,8 @@ func (c *Coin) ConfigureDaemon() daemon.Config {
 
 	dc.Pool.DefaultConnections = c.config.Node.DefaultConnections
 	dc.Pool.MaxDefaultPeerOutgoingConnections = c.config.Node.MaxDefaultPeerOutgoingConnections
+	dc.Pool.MaxIncomingMessageLength = c.config.Node.MaxIncomingMessageLength
+	dc.Pool.MaxOutgoingMessageLength = c.config.Node.MaxOutgoingMessageLength
 
 	dc.Pex.DataDirectory = c.config.Node.DataDirectory
 	dc.Pex.Disabled = c.config.Node.DisablePEX
@@ -359,6 +361,9 @@ func (c *Coin) ConfigureDaemon() daemon.Config {
 	dc.Pex.CustomPeersFile = c.config.Node.CustomPeersFile
 	dc.Pex.DefaultConnections = c.config.Node.DefaultConnections
 
+	dc.Daemon.MaxOutgoingMessageLength = uint64(c.config.Node.MaxOutgoingMessageLength)
+	dc.Daemon.MaxIncomingMessageLength = uint64(c.config.Node.MaxIncomingMessageLength)
+	dc.Daemon.MaxBlockTransactionsSize = c.config.Node.MaxBlockTransactionsSize
 	dc.Daemon.DefaultConnections = c.config.Node.DefaultConnections
 	dc.Daemon.DisableOutgoingConnections = c.config.Node.DisableOutgoingConnections
 	dc.Daemon.DisableIncomingConnections = c.config.Node.DisableIncomingConnections
@@ -385,7 +390,7 @@ func (c *Coin) ConfigureDaemon() daemon.Config {
 
 	dc.Visor.UnconfirmedVerifyTxn = c.config.Node.UnconfirmedVerifyTxn
 	dc.Visor.CreateBlockVerifyTxn = c.config.Node.CreateBlockVerifyTxn
-	dc.Visor.MaxBlockSize = c.config.Node.MaxBlockSize
+	dc.Visor.MaxBlockTransactionsSize = c.config.Node.MaxBlockTransactionsSize
 
 	dc.Visor.GenesisAddress = c.config.Node.genesisAddress
 	dc.Visor.GenesisSignature = c.config.Node.genesisSignature