+++ CHANGELOG.md
@@ -14,11 +14,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `POST /api/v2/wallet/transaction/sign` to sign an unsigned transaction with a wallet
 - Add `unsigned` option to `POST /api/v2/transaction/verify` for verifying an unsigned transaction
 - Add `POST /api/v2/transaction` to create an unsigned transaction from addresses or unspent outputs without a wallet
+- Add `-max-inc-msg-len` and `-max-out-msg-len` options to control the size of incoming and outgoing wire messages
 
 ### Fixed
 
 - Return v2-style error for disabled endpoints
 - #2172 Fix electron build failure for linux system
+- Don't send messages that exceed the configured 256kB limit, which caused peers to disconnect from the sender
 
 ### Changed
 
@@ -27,6 +29,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Use [`skyencoder`](https://github.com/skycoin/skyencoder)-generated binary encoders/decoders for network and database data, instead of the reflect-based encoders/decoders in `cipher/encoder`.
 - Add `/api/v1/resendUnconfirmedTxns` to the `WALLET` API set
 - In `POST /api/v1/wallet/transaction`, moved `wallet` parameters to the top level of the object
+- Incoming wire message size limit increased to 1024kB
+- Clients restrict the maximum number of blocks they will send in a `GiveBlocksMessage` to 20
 
 ### Removed